reorganize gen_py
