+++ try/python/for1.py
@@ -1,0 +1,1 @@
+for i in range(1, 5) print(i)
